Fix parsing errors.
The default HTML parser used by BS4 nests unclosed <td>-tags.
<td>1   <td>2   <td>3
becomes
<td>1   <td>2   <td>3   </td></td></td>

Then, the fields-list will contain as first td-element
<td>1   <td>2   <td>3   </td></td></td>
as second td-element
<td>2   <td>3   </td></td>
and as third td-element
 <td>3   </td>

When get_text() is called on the first element, this will provide all text
contained within, so it will become
"1   2   3   "

This seems to be a bug in the HTML parser, since td-tags cannot be nested.
This patch fixes it by using the html5lib parser. It does, however, require
users to install html5lib.